Move IsExactArgs and FindIPv6 to helpers
