+++ cmd/nerdctl/container_cp_linux.go
@@ -44,7 +44,7 @@ Using 'nerdctl cp' with untrusted or malicious containers is unsupported and may
   nerdctl cp [flags] SRC_PATH|- CONTAINER:DEST_PATH`
 	var cpCommand = &cobra.Command{
 		Use:               usage,
-		Args:              IsExactArgs(2),
+		Args:              helpers.IsExactArgs(2),
 		Short:             shortHelp,
 		Long:              longHelp,
 		RunE:              cpAction,
