+++ cmd/nerdctl/network_create_linux_test.go
@@ -19,11 +19,11 @@ package main
 import (
 	"fmt"
 	"net"
-	"strings"
 	"testing"
 
 	"gotest.tools/v3/assert"
 
+	"github.com/containerd/nerdctl/v2/cmd/nerdctl/helpers"
 	"github.com/containerd/nerdctl/v2/pkg/testutil"
 )
 
@@ -73,25 +73,10 @@ func TestNetworkCreateIPv6(t *testing.T) {
 	})
 
 	base.Cmd("run", "--rm", "--net", testNetwork, testutil.CommonImage, "ip", "addr", "show", "dev", "eth0").AssertOutWithFunc(func(stdout string) error {
-		ip := findIPv6(stdout)
+		ip := helpers.FindIPv6(stdout)
 		if subnet.Contains(ip) {
 			return nil
 		}
 		return fmt.Errorf("expected subnet %s include ip %s", subnet, ip)
 	})
-}
-
-func findIPv6(output string) net.IP {
-	var ipv6 string
-	lines := strings.Split(output, "\n")
-	for _, line := range lines {
-		if strings.Contains(line, "inet6") {
-			fields := strings.Fields(line)
-			if len(fields) > 1 {
-				ipv6 = strings.Split(fields[1], "/")[0]
-				break
-			}
-		}
-	}
-	return net.ParseIP(ipv6)
 }
